feat: add dao for playlist/mylist page and move dto
